requirements.txt
Flask == 0.11.1
Keras == 2.0.5
PyJWT == 1.4.2
dateparser == 0.6.0
numpy == 1.13.0
scipy == 0.18.0